review comment incorporation code changes
