+++ login-workflow/example/src/screens/Home.css
@@ -1,3 +1,7 @@
+.App {
+    text-align: center;
+}
+
 .App-logo {
     height: 40vmin;
     pointer-events: none;
