+++ login-workflow/src/new-architecure/components/workflow/WorkflowCardHeader.tsx
@@ -1,28 +1,25 @@
 import React from 'react';
 import CardHeader from '@mui/material/CardHeader';
 import Typography from '@mui/material/Typography';
-import { useLanguageLocale } from '../../../auth-shared';
-import { SxProps, Theme, useTheme } from '@mui/material/styles';
-
-export type WorkflowCardHeaderProps = {
-    title?: string;
-};
+import { useTheme } from '@mui/material/styles';
+import { WorkflowCardHeaderProps } from './WorkflowCardTypes';
 
 export const WorkflowCardHeader: React.FC<WorkflowCardHeaderProps> = (props) => {
-    const { t } = useLanguageLocale();
+    const { title, ...cardHeaderProps } = props;
     const theme = useTheme();
-    const { title } = props;
-
-    const DialogTitleStyles = (theme: Theme): SxProps<Theme> => ({
-        p: `${theme.spacing(4)} ${theme.spacing(3)} 0 ${theme.spacing(3)}`,
-        [theme.breakpoints.down('sm')]: {
-            p: `${theme.spacing(2)} ${theme.spacing(2)} 0 ${theme.spacing(2)}`,
-        },
-    });
 
     return (
         <>
-            <CardHeader title={<Typography variant={'h6'}>{t(title)}</Typography>} sx={DialogTitleStyles(theme)} />
+            <CardHeader
+                title={<Typography variant={'h6'}>{title}</Typography>}
+                sx={{
+                    p: `${theme.spacing(4)} ${theme.spacing(3)} 0 ${theme.spacing(3)}`,
+                    [theme.breakpoints.down('sm')]: {
+                        p: `${theme.spacing(2)} ${theme.spacing(2)} 0 ${theme.spacing(2)}`,
+                    },
+                }}
+                {...cardHeaderProps}
+            />
         </>
     );
 };
